feature(fix): refactor code - move dataFetching logic to their files [#167561791]
