+++ package.json
@@ -30,7 +30,6 @@
     "axios": "^0.19.0",
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
-    "google-auth-library": "^5.1.1",
     "graphql": "^14.4.2",
     "jsonwebtoken": "^8.5.1",
     "moment": "^2.24.0",
