+++ src/helpers/authentication.js
